+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix bug accessing statistics when not initialized 
+
 ## [1.6.0] - 2022-06-30
 
 ### Added